Add failing test
